refactor(microsoft-defender-o365): generalize source module filenames (#493)
