+++ microsoft-defender-o365/src/__main__.py
@@ -1,6 +1,6 @@
 """Main entry point for the collector."""
 
-from src.microsoft_defender_o365_collector import main
+from src.collector_main import main
 
 if __name__ == "__main__":
     main()
