+++ microsoft-defender-o365/src/collector_main.py
@@ -12,13 +12,9 @@ import sys
 
 from src.collector.collector import BaseCollector
 from src.collector.models.source import Source
-from src.source.microsoft_defender_o365_data_fetcher import (
-    MicrosoftDefenderO365DataFetcher,
-)
-from src.source.microsoft_defender_o365_signatures import SUPPORTED_SIGNATURES
-from src.source.microsoft_defender_o365_source_data import (
-    MicrosoftDefenderO365SourceData,
-)
+from src.source.data_fetcher import MicrosoftDefenderO365DataFetcher
+from src.source.signatures import SUPPORTED_SIGNATURES
+from src.source.source_data import MicrosoftDefenderO365SourceData
 
 
 def main() -> None:
