+++ microsoft-defender-o365/src/source/signatures.py
